+++ packages/substrate/src/wanxiang_substrate/rc001/__init__.py
@@ -9,9 +9,23 @@ from wanxiang_substrate.rc001.instantiate import (
     record_lineage_root,
     resolve_rc001_profile,
 )
+from wanxiang_substrate.rc001.npc import (
+    Activity,
+    ActivityKind,
+    NPCProfile,
+    daily_schedule,
+    resolve_population,
+    resolve_population_result,
+)
 
 __all__ = [
+    "Activity",
+    "ActivityKind",
     "InitialSnapshot",
+    "NPCProfile",
+    "daily_schedule",
+    "resolve_population",
+    "resolve_population_result",
     "RC001Instance",
     "RC001Profile",
     "genesis_delta",